+++ recipe-searcher/src/Components/FavRecipeCard.js
@@ -1,15 +1,11 @@
 import Cookies from "js-cookie";
 import { Link } from "react-router-dom";
 
-
 function FavRecipeCard( { setFav, name, id, image, description } ) {
-
-
 
     function handleDelete() {
         const userId = Cookies.get('flavorful_id')
     
-        
         setFav(prevRecipes => {
             const updatedRecipes = prevRecipes.filter(recipe => recipe.id !== id);
             
@@ -26,22 +22,15 @@ function FavRecipeCard( { setFav, name, id, image, description } ) {
         });
     }
     
-
-
     return (
         <div className="recipe-card">
             <h1>{name}</h1>
             <Link to={`/favoriterecipe/${id}`}>
             <img src={image} alt={name}/>
             </Link>
-            <div>
-                {description}
-            </div>
-            <div>
-                &nbsp;&nbsp;&nbsp;<button className="del-btn" onClick={() => handleDelete(id)}>Delete &#128465;</button>
-            </div>
+            <div>{description}</div>
+            <div>&nbsp;&nbsp;&nbsp;<button className="del-btn" onClick={() => handleDelete(id)}>Delete &#128465;</button></div>
         </div>
-        
     )};
 
     export default FavRecipeCard;
