+++ recipe-searcher/src/Components/MyRecipesCard.js
@@ -2,12 +2,8 @@ import Cookies from "js-cookie"
 import { Link } from "react-router-dom";
 
 function MyRecipesCard( {name, image, id, description, setMyRecipes}) {
-    
-    
-    
     function handleDelete() {
         const userId = Cookies.get('flavorful_id')
-    
         
         setMyRecipes(prevRecipes => {
             const updatedRecipes = prevRecipes.filter(recipe => recipe.id !== id);
@@ -25,20 +21,14 @@ function MyRecipesCard( {name, image, id, description, setMyRecipes}) {
         });
     }
     
-
-
 return (
     <div className="recipe-card">
-        <h1 >{name}</h1>
+        <h1>{name}</h1>
         <Link to={`/myrecipe/${id}`}>
         <img src={image} alt={name}/>
         </Link>
-        <div>
-            {description}
-        </div>
-        <div>
-            &nbsp;&nbsp;&nbsp;<button className="del-btn" onClick={() => handleDelete(id)}>Delete &#128465;</button>
-        </div>
+        <div>{description}</div>
+        <div>&nbsp;&nbsp;&nbsp;<button className="del-btn" onClick={() => handleDelete(id)}>Delete &#128465;</button></div>
     </div>
     
 )};
