+++ recipe-searcher/src/Components/Nav.js
@@ -3,7 +3,6 @@ import { Link } from "react-router-dom"
 function Nav() {
     return (
     <nav>
-        
         <Link to="/">Home</Link> &#160; &#160;
         <Link to="/myrecipes">My Recipes</Link>&#160; &#160;
         <Link to="/favorites" >Favorites</Link>&#160; &#160;
