+++ recipe-searcher/src/Pages/Favorites.js
@@ -1,5 +1,5 @@
 import { useEffect } from "react"
-import FavRecipeCard from "./FavRecipeCard"
+import FavRecipeCard from "../Components/FavRecipeCard"
 import Cookies from "js-cookie"
 
 function Favorites({ fav, setFav }) {
@@ -16,9 +16,7 @@ function Favorites({ fav, setFav }) {
  
     return (
         <div>
-            <div className="recipe-container">
-                {likedRecipes}
-            </div>
+            <div className="recipe-container">{likedRecipes}</div>
         </div>
         )
 }
