+++ recipe-searcher/src/Pages/RecipeDetails.js
@@ -1,9 +1,9 @@
 import { useParams } from "react-router-dom";
 
-function MyRecipeDetails({ myRecipes }) {
+function RecipeDetails({ recipes }) {
 
     const { id } = useParams();
-    const recipe = myRecipes.find(r => r.id === Number(id));
+    const recipe = recipes.find(r => r.id === id);
 
     if (!recipe) {
         return <div>Loading...</div>;
@@ -19,15 +19,11 @@ const stepsList = recipe.steps.map(step => {
 
 return(
     <div className="details">
-        <img alt={recipe.name} src={recipe.image} />
-        <h1>{recipe.name}</h1>
-    <div>
-        {ingredientsList}
-    </div>
-    <div>
-        {stepsList}
-    </div>
+    <img alt={recipe.name} src={recipe.image} />
+    <h1>{recipe.name}</h1>
+    <div>{ingredientsList}</div>
+    <div>{stepsList}</div>
     </div>
 )};
 
-export default MyRecipeDetails;
+export default RecipeDetails;
